created separate modules for dictionary and database functions
